Improve code quality and developer experience as a user of the package
- Use `Path` instead of `str` for path variables
- Fix pyarrow datatype for long integer: `use pa.uint64` instead of `pa.int32` and `pa.int64`
- Encode projection business rules into `ProjectionRef` type to clarify usage
- Add validation to EventSpec
- Extract output directories/files logic from `ShardWriter` into `ShardsDir` to re-use it outside and to favor dependency injection
- Replace some raw `tuple` and `dict` returns with dataclasses to clarify meaning
- Add more outputs to `fetch_decode` to ease package usage in a data flow
- Fix existing tests
- Add more tests
- Add examples and update the README

+++ src/defind/core/models.py
@@ -25,10 +25,10 @@ import pyarrow as pa
 # === Base schema (Arrow) ===
 
 _BASE_FIELDS: list[tuple[str, pa.DataType]] = [
-    ("block_number", pa.int64()),
-    ("block_timestamp", pa.int64()),
+    ("block_number", pa.uint64()),
+    ("block_timestamp", pa.uint64()),
     ("tx_hash", pa.string()),
-    ("log_index", pa.int32()),
+    ("log_index", pa.uint64()),
     ("contract", pa.string()),
     ("event", pa.string()),
 ]
@@ -214,10 +214,10 @@ class Column:
         """Convert the buffer to a sorted Arrow table with deterministic schema."""
         fields = [pa.field(n, t) for n, t in _BASE_FIELDS]
         arrays: dict[str, pa.Array] = {
-            "block_number": pa.array(self.block_number, type=pa.int64()),
-            "block_timestamp": pa.array(self.block_timestamp, type=pa.int64()),
+            "block_number": pa.array(self.block_number, type=pa.uint64()),
+            "block_timestamp": pa.array(self.block_timestamp, type=pa.uint64()),
             "tx_hash": pa.array(self.tx_hash, type=pa.string()),
-            "log_index": pa.array(self.log_index, type=pa.int32()),
+            "log_index": pa.array(self.log_index, type=pa.uint64()),
             "contract": pa.array(self.contract, type=pa.string()),
             "event": pa.array(self.event, type=pa.string()),
         }
